# src/api/openapi-client.json
{
  "openapi": "3.0.0",
  "security": [{ "ApiKey": [] }],
  "info": {
    "title": "Companion API",
    "version": "0.0.1"
  },
  "paths": {
    "/messages": {
      "post": {
        "summary": "Create a Message",
        "operationId": "createMessage",
        "tags": ["messages"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companionId": {
                    "type": "string",
                    "description": "The id of the companion"
                  },
                  "role": {
                    "type": "string",
                    "description": "The message role"
                  },
                  "content": {
                    "type": "string",
                    "description": "The message content"
                  }
                },
                "required": ["companionId", "content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": {
                      "type": "string",
                      "description": "The ID of the created message"
                    },
                    "role": {
                      "type": "string",
                      "description": "The message role"
                    },
                    "content": {
                      "type": "string",
                      "description": "Content of the message"
                    }
                  },
                  "required": ["messageId", "content"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        },
        "x-order": 0
      }
    },
    "/messages/{messageId}": {
      "get": {
        "summary": "Get a specific message",
        "operationId": "getMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 40 }
          }
        ],
        "tags": ["messages"],
        "responses": {
          "200": {
            "description": "Successfully fetched details about a message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": { "type": "string" },
                    "content": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2021-01-01T08:00:00Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2021-01-01T08:00:00Z"
                    }
                  },
                  "required": ["messageId", "content", "createdAt", "updatedAt"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific message",
        "operationId": "updateMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "content": { "type": "string" } },
                "required": ["content"]
              }
            }
          }
        },
        "tags": ["messages"],
        "responses": {
          "200": {
            "description": "Successfully updated details about a message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": { "type": "string" },
                    "content": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2021-01-01T08:00:00Z"
                    }
                  },
                  "required": ["messageId", "content", "updatedAt"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete a specific message",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["messages"],
        "responses": {
          "200": {
            "description": "Successfully delete a message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": { "type": "string" },
                    "content": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2021-01-01T08:00:00Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2021-01-01T08:00:00Z"
                    }
                  },
                  "required": ["messageId", "content", "createdAt", "updatedAt"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/messages/bulk": {
      "post": {
        "summary": "Create a bulk of messages",
        "operationId": "createBulkMessage",
        "tags": ["messages"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companionId": {
                    "type": "string",
                    "description": "The id of the companion"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "description": "The message role"
                        },
                        "content": {
                          "type": "string",
                          "description": "The message content"
                        }
                      },
                      "required": ["role", "content"]
                    }
                  }
                },
                "required": ["companionId", "messages"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "messageId": {
                        "type": "string",
                        "description": "The ID of the created message"
                      },
                      "content": {
                        "type": "string",
                        "description": "Content of the message"
                      },
                      "role": {
                        "type": "string",
                        "description": "Role of the message (e.g., 'user', 'assistant')"
                      }
                    },
                    "required": ["messageId", "content", "role"]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        },
        "x-order": 0
      }
    },
    "/companions": {
      "post": {
        "summary": "Create a Companion",
        "operationId": "createCompanion",
        "tags": ["companions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "the name of the companion"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a Companion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companionId": {
                      "type": "string",
                      "description": "The ID of the created Companion"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the companion"
                    }
                  },
                  "required": ["companionId", "name"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        },
        "x-order": 0
      },
      "get": {
        "summary": "Get a list of all companions",
        "operationId": "getCompanions",
        "parameters": [
          {
            "description": "Maximum number of results returned (25 by default)",
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 1000,
              "default": 25
            }
          },
          {
            "description": "Page of results to be fetched from the server (1 by default)",
            "name": "page",
            "in": "query",
            "schema": { "type": "number", "minimum": 1, "default": 1 }
          }
        ],
        "tags": ["companions"],
        "responses": {
          "200": {
            "description": "Successfully fetched a list of all Companions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "companionId": { "type": "string" },
                          "name": { "type": "string" },
                          "createdAt": { "type": "string" },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "companionId",
                          "name",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "description": "Information about the results",
                      "properties": {
                        "pageSize": {
                          "type": "number",
                          "description": "Maximum number of results return in a single request"
                        },
                        "previousPage": {
                          "type": "number",
                          "description": "Number of the previous page. If empty, means the current page is the first one"
                        },
                        "nextPage": {
                          "type": "number",
                          "description": "Number of the next page. If empty, means the current page is the last one"
                        },
                        "pageCount": {
                          "type": "number",
                          "description": "Total number of pages found"
                        },
                        "itemCount": {
                          "type": "number",
                          "description": "Total number of items found"
                        }
                      },
                      "required": ["pageSize", "pageCount", "itemCount"]
                    },
                    "links": {
                      "type": "object",
                      "properties": {
                        "self": { "type": "string" },
                        "next": { "type": "string" },
                        "prev": { "type": "string" }
                      },
                      "required": ["self"]
                    }
                  },
                  "required": ["items", "meta", "links"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
    "/companions/{companionId}": {
      "get": {
        "summary": "Get a companion",
        "operationId": "getCompanion",
        "parameters": [
          {
            "name": "companionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["companions"],
        "responses": {
          "200": {
            "description": "Successfully fetched a companion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companionId": {
                      "type": "string",

                      "description": "the id of the companion"
                    },
                    "name": {
                      "type": "string",
                      "description": "the name of the companion"
                    }
                  },
                  "required": ["companionId", "name"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
    "/companions/{companionId}/messages": {
      "get": {
        "summary": "Get a list of all messages related with a companion",
        "operationId": "getMessagesByCompanion",
        "parameters": [
          {
            "name": "companionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Maximum number of results returned (25 by default)",
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 1000,
              "default": 25
            }
          },
          {
            "description": "Page of results to be fetched from the server (1 by default)",
            "name": "page",
            "in": "query",
            "schema": { "type": "number", "minimum": 1, "default": 1 }
          }
        ],
        "tags": ["companions"],
        "responses": {
          "200": {
            "description": "Successfully fetched a list of all Messages associated with a Companion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "messageId": { "type": "string" },
                          "content": { "type": "string" },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "messageId",
                          "content",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "description": "Information about the results",
                      "properties": {
                        "pageSize": {
                          "type": "number",
                          "description": "Maximum number of results return in a single request"
                        },
                        "previousPage": {
                          "type": "number",
                          "description": "Number of the previous page. If empty, means the current page is the first one"
                        },
                        "nextPage": {
                          "type": "number",
                          "description": "Number of the next page. If empty, means the current page is the last one"
                        },
                        "pageCount": {
                          "type": "number",
                          "description": "Total number of pages found"
                        },
                        "itemCount": {
                          "type": "number",
                          "description": "Total number of items found"
                        }
                      },
                      "required": ["pageSize", "pageCount", "itemCount"]
                    },
                    "links": {
                      "type": "object",
                      "properties": {
                        "self": { "type": "string" },
                        "next": { "type": "string" },
                        "prev": { "type": "string" }
                      },
                      "required": ["self"]
                    }
                  },
                  "required": ["items", "meta", "links"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": { "type": "apiKey", "name": "apiKey", "in": "header" }
    },
...
  },
...
}